Keep Double data type for "start" field
Before https://github.com/cdsap/Talaiot/pull/160 "start" field was exported as Double data type. This PR changes data type back from Long to Double. Otherwise InfluxDB publisher fails with "field type conflict" on existing database.

Add other missing fields from ExecutionReport to the test.

Fixes: https://github.com/cdsap/Talaiot/issues/186

+++ talaiot/src/main/kotlin/com/cdsap/talaiot/metrics/DefaultBuildDataProvider.kt
@@ -18,7 +18,7 @@ class DefaultBuildMetricsProvider(
             buildInvocationId?.let { map["buildInvocationId"] = it }
             requestedTasks?.let { map["requestedTasks"] = it }
             cacheRatio?.let { map["cacheRatio"] = it.toDouble() }
-            beginMs?.let { map["start"] = it.toLong() }
+            beginMs?.let { map["start"] = it.toDouble() }
             rootProject?.let { map["rootProject"] = it }
             scanLink?.let { map["scanLink"] = it }
